deterministic-enterprise-ai-agent/requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langgraph
langchain-community
langchain-google-genai>=2.0.0
langchain-google-vertexai
langchain-text-splitters
langchain-huggingface

faiss-cpu

# Modern Python 3.13 compatible stack
sentence-transformers==5.2.0
transformers==4.57.5
torch==2.9.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

huggingface-hub
pydantic

# Vertex AI SDKs
google-cloud-aiplatform